21.4 - showtasks.bat

+++ src/main/java/com/crud/tasks/service/DbService.java
@@ -27,8 +27,12 @@ public class DbService {
         return repository.save(task);
     }
 
-    public void deleteTask(final Long taskId) {
-        repository.deleteById(taskId);
+    public void deleteTask(final Long taskId) throws TaskNotFoundException {
+        if(repository.existsById(taskId)) {
+            repository.deleteById(taskId);
+        } else {
+            throw new TaskNotFoundException();
+        }
     }
 
 }
